Update hooks for VulnerabilityScan.tsx
