+++ frontend/src/pages/VulnerabilityScanDash/VulnerabilityScan.tsx
@@ -1,11 +1,13 @@
-import React, { useState, useEffect, useMemo } from 'react';
+import React, { useMemo } from 'react';
 import { withSearch } from '@elastic/react-search-ui';
 import { Box, BoxProps, Grid, CircularProgress, Link } from '@mui/material';
 // Context & hooks
 import { ContextType, useAuthContext } from 'context';
 import { useVulnScanData } from 'hooks/useVulnScanData';
 import { useCheckUserState } from 'hooks/useCheckUserState';
-import { useMaintenanceNotifications } from 'hooks/useMaintenanceNotifications';
+import { useMaintenanceNotifications } from '@/hooks/useMaintenanceNotifications';
+import { useClearFiltersOnMount } from 'hooks/useClearFiltersOnMount';
+import { useOrgInfo } from 'hooks/useOrgInfo';
 // Constants & types
 import { ORGANIZATION_FILTER_KEY } from 'components/FilterDrawer/VSDashRegionAndOrgFilters';
 import { OrganizationShallow } from 'components/FilterDrawer/RegionAndOrganizationFilters';
@@ -79,32 +81,9 @@ export const VulnerabilityScan: React.FC<ContextType> = (props) => {
     setMaintenanceNotification
   } = useMaintenanceNotifications(user, apiGet, userMustSign);
 
-  useEffect(() => {
-    filters.forEach((filter) => {
-      removeFilter(filter.field, filter.values[0], filter.type);
-    });
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  useClearFiltersOnMount(filters, removeFilter);
 
-  const orgId = useMemo(() => {
-    const organizationFilter = filters.find(
-      (filter) => filter.field === ORGANIZATION_FILTER_KEY
-    );
-    return (
-      organizationFilter?.values[0]?.id ||
-      organizationFilter?.values[0] ||
-      currentOrganization?.id ||
-      (currentOrganization as OrganizationShallow)?.id
-    );
-  }, [filters, currentOrganization]);
-
-  const organizationFilter = filters.find(
-    (filter) => filter.field === ORGANIZATION_FILTER_KEY
-  );
-
-  const orgName = organizationFilter?.values[0]?.name
-    ? organizationFilter?.values[0]?.name
-    : currentOrganization?.name;
+  const { orgId, orgName } = useOrgInfo(filters, currentOrganization);
 
   const {
     data: vulnScanData,
